processor/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.4
openpyxl==3.1.2
pytesseract==0.3.10
Pillow==10.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
